create TransorfmerRasaModel
